Se realizó documentación del controlador CMayorista.

+++ src/inventario/modelo/Mayorista.java
@@ -254,7 +254,7 @@ public class Mayorista {
         return diaActual;
     }
 
-    public double getInventario() {
+    public int getInventario() {
         return inventario;
     }
